+++ SinoFoundation/src/main/java/games/moegirl/sinocraft/sinofoundation/block/WoodChairsBlock.java
@@ -1,11 +1,16 @@
 package games.moegirl.sinocraft.sinofoundation.block;
 
+import games.moegirl.sinocraft.sinocore.capability.SCCapabilities;
+import games.moegirl.sinocraft.sinocore.capability.entity.armor_stand.SummonedArmorStand;
 import games.moegirl.sinocraft.sinocore.utility.ArmorStandHelper;
-import games.moegirl.sinocraft.sinocore.utility.VoxelShapeHelper;
+import games.moegirl.sinocraft.sinocore.utility.shape.VoxelShapeHelper;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.decoration.ArmorStand;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
