+++ SinoFoundation/src/main/java/games/moegirl/sinocraft/sinofoundation/block/WoodDeskBlock.java
@@ -1,6 +1,6 @@
 package games.moegirl.sinocraft.sinofoundation.block;
 
-import games.moegirl.sinocraft.sinocore.utility.VoxelShapeHelper;
+import games.moegirl.sinocraft.sinocore.utility.shape.VoxelShapeHelper;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.item.context.BlockPlaceContext;
